Disable a bunch of code to get program in a more compilable state

+++ ecs_damage.cpp
@@ -10,7 +10,7 @@ namespace ecs
 {
 	std::unordered_set<entt::entity> _entities_that_took_damage;
 
-	bool apply_damage_in_box(const Damage& damage, const Vector2f& box_min, const Vector2f& box_max, uint16_t mask_bits)
+	bool apply_damage_in_box(const Damage& damage, const Vector2f& box_min, const Vector2f& box_max, uint32_t mask_bits)
 	{
 		//debug::draw_box(box_min, box_max, Color::Red, 0.2f);
 		for (const OverlapHit& hit : overlap_box(box_min, box_max, mask_bits)) {
@@ -24,7 +24,7 @@ namespace ecs
 		return any_damaged;
 	}
 
-	bool apply_damage_in_circle(const Damage& damage, const Vector2f& center, float radius, uint16_t mask_bits)
+	bool apply_damage_in_circle(const Damage& damage, const Vector2f& center, float radius, uint32_t mask_bits)
 	{
 		//debug::draw_circle(center, radius, Color::Red, 0.2f);
 		for (const OverlapHit& hit : overlap_circle(center, radius, mask_bits)) {
